Naruto word guess
This name focuses on the user having to guess the randomly generated word.
You can create a list from which the word would have to be guessed and also
set a cap on the number of guesses allowed. After this, you can create
the rules yourself! When the user inputs the word, you can indicate whether
the alphabet written appears in this particular position or not. You will need
a function to check if the user is inputting alphabets or numbers and to
display error messages appropriately.